pyproject.toml
[project]
name = "character-tavern-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
]
